Day4 - Customerdetails
Create a Class called customer with properties customername, customerId  and with two methods – place order , Getcustomertimezone . Where the GetCustomerTimeZone  method is static.  Instantiate the class and call both the methods.  Copy the object to a new variable and change one of the properties.  And display the value of both objects.