Day4-StatesConditionalStatements
Accept State and Country name from User,  Check if country name is "India"  and if the condition is met  check if state is "Kerala" , then rerun user true  or false.  Do this using conditional / ternary operators.  the names entered by user should be case insensitive ("INDIA" and "India" are the same) , use this in your conditional checks.