July13 - Employee details
Design an  Employee class with properties EmployeeId, Name, Designation. Have an encapsulated field called Salary. Expose method called GetSalary() which displays or returns the current salary,  but setting of salary should be done at the time of hiring (object creation time).  Have an encapsulated method called IncreasePay() , where the salary can be increased by 10%. The user / client  should be exposed only to a method called Promote() with a boolean value, which decides the Increase Pay of 10% or not.