July13-Studentprogram
Write a program which has a class of Student with Id, Name and Date of Birth.  You can set the Date of birth at the time of Admission (object creation),  and have a property which calculates the age.  (use property  get accessors ) and can be accessed from the client / main program. Also ensure user can not enter the Dob or Age at a later point of time in the program